requirements.txt
pandas==2.0.0
numpy==1.23.5
scikit-learn==1.0.2
tensorflow==2.12.0
nltk==3.8.1
spacy==3.5.3
transformers==4.28.1
torch==2.0.0
fastapi==0.95.1
uvicorn==0.22.0
pymongo==4.3.3
python-dotenv==1.0.0
matplotlib==3.7.1
seaborn==0.12.2
requests==2.28.2
tqdm==4.65.0
pillow==9.5.0
opencv-python==4.7.0.72
python-multipart==0.0.6
torchvision==0.15.1
albumentations==1.3.0
efficientnet-pytorch==0.7.1
tensorboard==2.12.3
googletrans==3.1.0a0
langdetect==1.0.9
SpeechRecognition==3.10.0
gTTS==2.3.2
pygame==2.5.0
PyAudio==0.2.13
deepseek-ai==0.1.0
aiohttp==3.8.5
tenacity==8.2.3